// apps/admin-web/src/.generated/api/model/termsControllerGetAdminTermsV1Params.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Admin API
 * Control Plane Admin API Service
 * OpenAPI spec version: 1.0.0
 */
import type { SortDirection } from './sortDirection';

export type TermsControllerGetAdminTermsV1Params = {
sort?: string[];
direction?: SortDirection[];
search?: string;
page?: number;
/**
 * @maximum 100
 */
limit?: number;
groupId?: string;
};